# Conflicts:
#	Drtaa-Android/feature-rent/build.gradle.kts

+++ Drtaa-Android/core-model/src/main/java/com/drtaa/core_model/network/RequestDrivingCar.kt
@@ -1,6 +1,6 @@
 package com.drtaa.core_model.network
 
 data class RequestDrivingCar(
-    val rentCarId: Int,
+    val rentCarId: Long,
     val rentCarDrivingStatus: String
 )
